feat: make GradingObserverImpl class
